+++ docs/release-notes.md
@@ -1,4 +1,5 @@
 # New features and Changes
 
 # Bug fixes
+- Fix anime with long titles being downloaded with the shortened title
 - senpcli: Fix tracking check tracked anime crashing if no episode of the anime was initially downloaded